GCA-102 Implement aux endpoints

+++ src/main/java/com/gym/crm/app/facade/GymFacade.java
@@ -8,7 +8,6 @@ import com.gym.crm.app.domain.dto.trainer.TrainerDto;
 import com.gym.crm.app.domain.dto.trainer.TrainerUpdateRequest;
 import com.gym.crm.app.domain.dto.training.TrainingDto;
 import com.gym.crm.app.domain.dto.training.TrainingSaveRequest;
-import com.gym.crm.app.domain.dto.user.UserCredentialsDto;
 import com.gym.crm.app.mapper.TraineeMapper;
 import com.gym.crm.app.mapper.TrainerMapper;
 import com.gym.crm.app.mapper.TrainingMapper;
@@ -16,6 +15,7 @@ import com.gym.crm.app.mapper.UserMapper;
 import com.gym.crm.app.repository.search.filters.TraineeTrainingSearchFilter;
 import com.gym.crm.app.repository.search.filters.TrainerTrainingSearchFilter;
 import com.gym.crm.app.rest.AvailableTrainerGetResponse;
+import com.gym.crm.app.rest.ChangePasswordRequest;
 import com.gym.crm.app.rest.TraineeAssignedTrainersUpdateRequest;
 import com.gym.crm.app.rest.TraineeAssignedTrainersUpdateResponse;
 import com.gym.crm.app.rest.TraineeCreateResponse;
@@ -31,6 +31,7 @@ import com.gym.crm.app.rest.TrainingWithTraineeName;
 import com.gym.crm.app.rest.TrainingWithTrainerName;
 import com.gym.crm.app.security.AuthenticationService;
 import com.gym.crm.app.security.CurrentUserHolder;
+import com.gym.crm.app.security.model.UserCredentialsDto;
 import com.gym.crm.app.service.TraineeService;
 import com.gym.crm.app.service.TrainerService;
 import com.gym.crm.app.service.TrainingService;
@@ -41,6 +42,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
+
+import static com.gym.crm.app.security.UserRole.ADMIN;
+import static com.gym.crm.app.security.UserRole.TRAINEE;
+import static com.gym.crm.app.security.UserRole.TRAINER;
 
 @Getter
 @Component
@@ -89,32 +94,32 @@ public class GymFacade {
     }
 
     public List<TrainerDto> getAllTrainers(UserCredentialsDto userCredentials) {
-        authService.authenticate(userCredentials);
+        authService.checkUserAuthorisation(userCredentials, ADMIN);
 
         return trainerService.getAllTrainers();
     }
 
     public List<TraineeDto> getAllTrainees(UserCredentialsDto userCredentials) {
-        authService.authenticate(userCredentials);
+        authService.checkUserAuthorisation(userCredentials, ADMIN);
 
         return traineeService.getAllTrainees();
     }
 
     public TrainerGetResponse getTrainerByUsername(String username) {
-        authService.authenticate(getCurrentCredentials());
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINER, TRAINEE);
 
         return trainerMapper.toGetResponse(trainerService.getTrainerByUsername(username));
     }
 
     public TraineeGetResponse getTraineeByUsername(String username) {
-        authService.authenticate(getCurrentCredentials());
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINER, TRAINEE);
 
         return traineeMapper.dtoToGetResponse(traineeService.getTraineeByUsername(username));
     }
 
     public TraineeAssignedTrainersUpdateResponse updateTraineeTrainersList(String username,
                                                                            TraineeAssignedTrainersUpdateRequest request) {
-        authService.authenticate(getCurrentCredentials());
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINER, TRAINEE);
 
         List<Trainer> trainers = traineeService.updateTraineeTrainersByUsername(username, request.getTrainerUsernames()).stream()
                 .map(trainerMapper::entityToRestTrainer).toList();
@@ -124,36 +129,36 @@ public class GymFacade {
 
     public TraineeUpdateResponse updateTraineeByUsername(String username,
                                                          @Valid TraineeUpdateRequest updateRequest) {
-        authService.authenticate(getCurrentCredentials());
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINEE);
 
         return traineeMapper.dtoToUpdateResponse(traineeService.updateTraineeByUsername(username, updateRequest));
     }
 
     public TrainerUpdateResponse updateTrainerByUsername(String username,
                                                          @Valid TrainerUpdateRequest updateRequest) {
-        authService.authenticate(getCurrentCredentials());
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINER);
 
         return trainerMapper.toUpdateResponse(trainerService.updateTrainerByUsername(username, updateRequest));
     }
 
     public void deleteTrainerByUsername(String username, UserCredentialsDto userCredentials) {
-        authService.authenticate(userCredentials);
+        authService.checkUserAuthorisation(userCredentials, ADMIN, TRAINER);
 
         trainerService.deleteTrainerByUsername(username);
     }
 
     public void deleteTraineeByUsername(String username) {
-        authService.authenticate(getCurrentCredentials());
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINEE);
 
         traineeService.deleteTraineeByUsername(username);
     }
 
     public TrainerTrainingGetResponse getTrainerTrainingsByFilter(@Valid TrainerTrainingSearchFilter criteria) {
-        authService.authenticate(getCurrentCredentials());
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINER);
 
         List<TrainingDto> trainings = trainingService.getTrainerTrainingsByFilter(criteria);
 
-        List<TrainingWithTraineeName>trainingWithTraineeNames = trainings.stream()
+        List<TrainingWithTraineeName> trainingWithTraineeNames = trainings.stream()
                 .map(this::buildTrainingWithTraineeName)
                 .toList();
 
@@ -161,7 +166,7 @@ public class GymFacade {
     }
 
     public TraineeTrainingGetResponse getTraineeTrainingsByFilter(@Valid TraineeTrainingSearchFilter filter) {
-        authService.authenticate(getCurrentCredentials());
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINER, TRAINEE);
 
         List<TrainingDto> trainings = trainingService.getTraineeTrainingsByFilter(filter);
 
@@ -173,7 +178,7 @@ public class GymFacade {
     }
 
     public AvailableTrainerGetResponse getUnassignedTrainersByTraineeUsername(String username) {
-        authService.authenticate(getCurrentCredentials());
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINEE);
 
         List<Trainer> trainers = traineeService.getUnassignedTrainersByTraineeUsername(username).stream()
                 .map(trainerMapper::toEntity).toList();
@@ -182,38 +187,37 @@ public class GymFacade {
     }
 
     public List<TrainingDto> getAllTrainings(UserCredentialsDto userCredentials) {
-        authService.authenticate(userCredentials);
+        authService.checkUserAuthorisation(userCredentials, ADMIN);
 
         return trainingService.getAllTrainings();
     }
 
     public TrainingDto addTraining(@Valid TrainingSaveRequest createRequest, UserCredentialsDto userCredentials) {
-        authService.authenticate(userCredentials);
+        authService.checkUserAuthorisation(userCredentials, ADMIN);
 
         return trainingService.addTraining(createRequest);
     }
 
     public TrainingDto updateTraining(@Valid TrainingSaveRequest updateRequest, UserCredentialsDto userCredentials) {
-        authService.authenticate(userCredentials);
+        authService.checkUserAuthorisation(userCredentials, ADMIN);
 
         return trainingService.updateTraining(updateRequest);
     }
 
     public void switchActivationStatus(String username) {
-        authService.authenticate(getCurrentCredentials());
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINER, TRAINEE);
 
         userProfileService.switchActivationStatus(username);
     }
 
-    public void changePassword(String username, String password, UserCredentialsDto userCredentials) {
-        authService.authenticate(userCredentials);
+    public void changePassword(ChangePasswordRequest request) {
+        authService.checkUserAuthorisation(getCurrentCredentials(), ADMIN, TRAINER, TRAINEE);
 
-        userProfileService.changePassword(username, password);
+        userProfileService.changePassword(request.getUsername(), request.getOldPassword(), request.getNewPassword());
     }
 
     private UserCredentialsDto getCurrentCredentials() {
-        return userMapper.toCredentialsDto(
-                currentUserHolder.get());
+        return userMapper.toCredentialsDto(currentUserHolder.get());
     }
 
     private TrainingWithTrainerName buildTrainingWithTrainerName(TrainingDto trainingDto) {
@@ -243,5 +247,4 @@ public class GymFacade {
 
         return trainingWithTraineeName;
     }
-
 }
